+++ server/src/main/java/org/apache/druid/server/BrokerQueryResource.java
@@ -26,14 +26,11 @@ import com.sun.jersey.spi.container.ResourceFilters;
 import org.apache.druid.client.ServerViewUtil;
 import org.apache.druid.client.TimelineServerView;
 import org.apache.druid.guice.annotations.Json;
-import org.apache.druid.guice.annotations.Self;
-import org.apache.druid.guice.annotations.Smile;
 import org.apache.druid.query.CloneQueryMode;
 import org.apache.druid.query.Query;
 import org.apache.druid.query.QueryContexts;
 import org.apache.druid.query.planning.ExecutionVertex;
 import org.apache.druid.server.http.security.StateResourceFilter;
-import org.apache.druid.server.security.AuthConfig;
 import org.apache.druid.server.security.AuthorizerMapper;
 
 import javax.annotation.Nullable;
@@ -61,24 +58,20 @@ public class BrokerQueryResource extends QueryResource
   public BrokerQueryResource(
       QueryLifecycleFactory queryLifecycleFactory,
       @Json ObjectMapper jsonMapper,
-      @Smile ObjectMapper smileMapper,
       QueryScheduler queryScheduler,
-      AuthConfig authConfig,
       AuthorizerMapper authorizerMapper,
-      ResponseContextConfig responseContextConfig,
-      @Self DruidNode selfNode,
+      QueryResourceQueryResultPusherFactory queryResultPusherFactory,
+      ResourceIOReaderWriterFactory resourceIOReaderWriterFactory,
       TimelineServerView brokerServerView
   )
   {
     super(
         queryLifecycleFactory,
         jsonMapper,
-        smileMapper,
         queryScheduler,
-        authConfig,
         authorizerMapper,
-        responseContextConfig,
-        selfNode
+        queryResultPusherFactory,
+        resourceIOReaderWriterFactory
     );
     this.brokerServerView = brokerServerView;
   }
@@ -96,7 +89,7 @@ public class BrokerQueryResource extends QueryResource
       @Context final HttpServletRequest req
   ) throws IOException
   {
-    final ResourceIOReaderWriter ioReaderWriter = createResourceIOReaderWriter(req, pretty != null);
+    final ResourceIOReaderWriterFactory.ResourceIOReaderWriter ioReaderWriter = resourceIOReaderWriterFactory.factorize(req, pretty != null);
     final CloneQueryMode cloneQueryMode = QueryContexts.getAsEnum(
         QueryContexts.CLONE_QUERY_MODE,
         cloneQueryModeString,
