+++ server/src/main/java/org/apache/druid/server/QueryResource.java
@@ -100,14 +100,11 @@ public class QueryResource implements QueryCountStatsProvider
 
   protected final QueryLifecycleFactory queryLifecycleFactory;
   protected final ObjectMapper jsonMapper;
-  protected final ObjectMapper smileMapper;
-  protected final ObjectMapper serializeDateTimeAsLongJsonMapper;
-  protected final ObjectMapper serializeDateTimeAsLongSmileMapper;
   protected final QueryScheduler queryScheduler;
   protected final AuthorizerMapper authorizerMapper;
 
-  private final ResponseContextConfig responseContextConfig;
-  private final DruidNode selfNode;
+  private final QueryResourceQueryResultPusherFactory queryResultPusherFactory;
+  protected final ResourceIOReaderWriterFactory resourceIOReaderWriterFactory;
 
   private final AtomicLong successfulQueryCount = new AtomicLong();
   private final AtomicLong failedQueryCount = new AtomicLong();
@@ -119,23 +116,18 @@ public class QueryResource implements QueryCountStatsProvider
   public QueryResource(
       QueryLifecycleFactory queryLifecycleFactory,
       @Json ObjectMapper jsonMapper,
-      @Smile ObjectMapper smileMapper,
       QueryScheduler queryScheduler,
-      AuthConfig authConfig,
       AuthorizerMapper authorizerMapper,
-      ResponseContextConfig responseContextConfig,
-      @Self DruidNode selfNode
+      QueryResourceQueryResultPusherFactory queryResultPusherFactory,
+      ResourceIOReaderWriterFactory resourceIOReaderWriterFactory
   )
   {
     this.queryLifecycleFactory = queryLifecycleFactory;
     this.jsonMapper = jsonMapper;
-    this.smileMapper = smileMapper;
-    this.serializeDateTimeAsLongJsonMapper = serializeDataTimeAsLong(jsonMapper);
-    this.serializeDateTimeAsLongSmileMapper = serializeDataTimeAsLong(smileMapper);
     this.queryScheduler = queryScheduler;
     this.authorizerMapper = authorizerMapper;
-    this.responseContextConfig = responseContextConfig;
-    this.selfNode = selfNode;
+    this.queryResultPusherFactory = queryResultPusherFactory;
+    this.resourceIOReaderWriterFactory = resourceIOReaderWriterFactory;
   }
 
   @DELETE
@@ -176,9 +168,7 @@ public class QueryResource implements QueryCountStatsProvider
       @Context final HttpServletRequest req
   ) throws IOException
   {
-    final QueryLifecycle queryLifecycle = queryLifecycleFactory.factorize();
-
-    final ResourceIOReaderWriter io = createResourceIOReaderWriter(req, pretty != null);
+    final ResourceIOReaderWriterFactory.ResourceIOReaderWriter io = resourceIOReaderWriterFactory.factorize(req, pretty != null);
 
     final String currThreadName = Thread.currentThread().getName();
     try {
@@ -190,6 +180,7 @@ public class QueryResource implements QueryCountStatsProvider
         return io.getResponseWriter().buildNonOkResponse(e.getFailType().getExpectedStatus(), e);
       }
 
+      final QueryLifecycle queryLifecycle = queryLifecycleFactory.factorize();
       queryLifecycle.initialize(query);
       final String queryThreadName = queryLifecycle.threadName(currThreadName);
       Thread.currentThread().setName(queryThreadName);
@@ -219,7 +210,8 @@ public class QueryResource implements QueryCountStatsProvider
         throw new ForbiddenException(authResult.getErrorMessage());
       }
 
-      final QueryResourceQueryResultPusher pusher = new QueryResourceQueryResultPusher(req, queryLifecycle, io);
+      final QueryResourceQueryResultPusherFactory.QueryResourceQueryResultPusher pusher =
+          queryResultPusherFactory.factorize(counter, req, queryLifecycle, io);
       return pusher.push();
     }
     catch (Exception e) {
@@ -273,7 +265,7 @@ public class QueryResource implements QueryCountStatsProvider
   private Query<?> readQuery(
       final HttpServletRequest req,
       final InputStream in,
-      final ResourceIOReaderWriter ioReaderWriter
+      final ResourceIOReaderWriterFactory.ResourceIOReaderWriter ioReaderWriter
   ) throws IOException
   {
     final Query<?> baseQuery;
@@ -301,120 +293,6 @@ public class QueryResource implements QueryCountStatsProvider
   private static String getPreviousEtag(final HttpServletRequest req)
   {
     return req.getHeader(HEADER_IF_NONE_MATCH);
-  }
-
-  protected ObjectMapper serializeDataTimeAsLong(ObjectMapper mapper)
-  {
-    return mapper.copy().registerModule(new SimpleModule().addSerializer(DateTime.class, new DateTimeSerializer()));
-  }
-
-  protected ResourceIOReaderWriter createResourceIOReaderWriter(HttpServletRequest req, boolean pretty)
-  {
-    String requestType = req.getContentType();
-    String acceptHeader = req.getHeader("Accept");
-
-    // response type defaults to Content-Type if 'Accept' header not provided
-    String responseType = Strings.isNullOrEmpty(acceptHeader) ? requestType : acceptHeader;
-
-    boolean isRequestSmile = SmileMediaTypes.APPLICATION_JACKSON_SMILE.equals(requestType) || APPLICATION_SMILE.equals(
-        requestType);
-    boolean isResponseSmile = SmileMediaTypes.APPLICATION_JACKSON_SMILE.equals(responseType)
-                              || APPLICATION_SMILE.equals(responseType);
-
-    return new ResourceIOReaderWriter(
-        isRequestSmile ? smileMapper : jsonMapper,
-        new ResourceIOWriter(
-            isResponseSmile ? SmileMediaTypes.APPLICATION_JACKSON_SMILE : MediaType.APPLICATION_JSON,
-            isResponseSmile ? smileMapper : jsonMapper,
-            isResponseSmile ? serializeDateTimeAsLongSmileMapper : serializeDateTimeAsLongJsonMapper,
-            pretty
-        )
-    );
-  }
-
-  protected static class ResourceIOReaderWriter
-  {
-    private final ObjectMapper requestMapper;
-    private final ResourceIOWriter writer;
-
-    public ResourceIOReaderWriter(ObjectMapper requestMapper, ResourceIOWriter writer)
-    {
-      this.requestMapper = requestMapper;
-      this.writer = writer;
-    }
-
-    public ObjectMapper getRequestMapper()
-    {
-      return requestMapper;
-    }
-
-    public ResourceIOWriter getResponseWriter()
-    {
-      return writer;
-    }
-  }
-
-  protected static class ResourceIOWriter
-  {
-    private final String responseType;
-    private final ObjectMapper inputMapper;
-    private final ObjectMapper serializeDateTimeAsLongInputMapper;
-    private final boolean isPretty;
-
-    ResourceIOWriter(
-        String responseType,
-        ObjectMapper inputMapper,
-        ObjectMapper serializeDateTimeAsLongInputMapper,
-        boolean isPretty
-    )
-    {
-      this.responseType = responseType;
-      this.inputMapper = inputMapper;
-      this.serializeDateTimeAsLongInputMapper = serializeDateTimeAsLongInputMapper;
-      this.isPretty = isPretty;
-    }
-
-    String getResponseType()
-    {
-      return responseType;
-    }
-
-    ObjectMapper newOutputWriter(
-        @Nullable QueryToolChest<?, Query<?>> toolChest,
-        @Nullable Query<?> query,
-        boolean serializeDateTimeAsLong
-    )
-    {
-      final ObjectMapper mapper = serializeDateTimeAsLong ? serializeDateTimeAsLongInputMapper : inputMapper;
-      final ObjectMapper decoratedMapper;
-      if (toolChest != null) {
-        decoratedMapper = toolChest.decorateObjectMapper(mapper, Preconditions.checkNotNull(query, "query"));
-      } else {
-        decoratedMapper = mapper;
-      }
-      return isPretty ? decoratedMapper.copy().enable(SerializationFeature.INDENT_OUTPUT) : decoratedMapper;
-    }
-
-    Response ok(Object object) throws IOException
-    {
-      return Response.ok(newOutputWriter(null, null, false).writeValueAsString(object), responseType).build();
-    }
-
-    Response gotError(Exception e) throws IOException
-    {
-      return buildNonOkResponse(
-          Status.INTERNAL_SERVER_ERROR.getStatusCode(),
-          QueryInterruptedException.wrapIfNeeded(e)
-      );
-    }
-
-    Response buildNonOkResponse(int status, Exception e) throws JsonProcessingException
-    {
-      return Response.status(status)
-                     .type(responseType)
-                     .entity(newOutputWriter(null, null, false).writeValueAsBytes(e))
-                     .build();
-    }
   }
 
   @Override
@@ -474,96 +352,6 @@ public class QueryResource implements QueryCountStatsProvider
     public void incrementTimedOut()
     {
       timedOutQueryCount.incrementAndGet();
-    }
-  }
-
-  private class QueryResourceQueryResultPusher extends QueryResultPusher
-  {
-    private final HttpServletRequest req;
-    private final QueryLifecycle queryLifecycle;
-    private final ResourceIOReaderWriter io;
-
-    public QueryResourceQueryResultPusher(
-        HttpServletRequest req,
-        QueryLifecycle queryLifecycle,
-        ResourceIOReaderWriter io
-    )
-    {
-      super(
-          req,
-          QueryResource.this.jsonMapper,
-          QueryResource.this.responseContextConfig,
-          QueryResource.this.selfNode,
-          QueryResource.this.counter,
-          queryLifecycle.getQueryId(),
-          MediaType.valueOf(io.getResponseWriter().getResponseType()),
-          ImmutableMap.of()
-      );
-      this.req = req;
-      this.queryLifecycle = queryLifecycle;
-      this.io = io;
-    }
-
-    @Override
-    public ResultsWriter start()
-    {
-      return new ResultsWriter()
-      {
-        private QueryResponse<Object> queryResponse;
-
-        @Override
-        public Response.ResponseBuilder start()
-        {
-          queryResponse = queryLifecycle.execute();
-          final ResponseContext responseContext = queryResponse.getResponseContext();
-          final String prevEtag = getPreviousEtag(req);
-
-          if (prevEtag != null && prevEtag.equals(responseContext.getEntityTag())) {
-            queryLifecycle.emitLogsAndMetrics(null, req.getRemoteAddr(), -1);
-            counter.incrementSuccess();
-            return Response.status(Status.NOT_MODIFIED);
-          }
-
-          return null;
-        }
-
-        @Override
-        public QueryResponse<Object> getQueryResponse()
-        {
-          return queryResponse;
-        }
-
-        @Override
-        public Writer makeWriter(OutputStream out) throws IOException
-        {
-          return new NativeQueryWriter(queryLifecycle.newOutputWriter(io), out);
-        }
-
-        @Override
-        public void recordSuccess(long numBytes)
-        {
-          queryLifecycle.emitLogsAndMetrics(null, req.getRemoteAddr(), numBytes);
-        }
-
-        @Override
-        public void recordFailure(Exception e)
-        {
-          queryLifecycle.emitLogsAndMetrics(e, req.getRemoteAddr(), -1);
-        }
-
-        @Override
-        public void close()
-        {
-
-        }
-      };
-    }
-
-    @Override
-    public void writeException(Exception e, OutputStream out) throws IOException
-    {
-      final ObjectMapper objectMapper = queryLifecycle.newOutputWriter(io);
-      out.write(objectMapper.writeValueAsBytes(e));
     }
   }
